Converted WydinFoodStore from kts to kt in order to add companion objects to be able to statically reference the attribute keys from another plugin
